requirements.txt
chess==1.10.0
pillow==10.4.0
pip==24.1.2
python-dotenv==1.0.1
